pyproject.toml
[tool.poetry]
name = "comoros"
version = "0.1.0"
description = "Wrapper package for GenomicSurveillance."
authors = ["Harald Vohringer"]

[tool.poetry.scripts]
comoros = "comoros.main:app"

[tool.poetry.dependencies]
python = "^3.7.1"
numpyro = "^0.5.0"
genomicsurveillance = "^0.4.1"
typer = {extras = ["all"], version = "^0.1.0"}

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-cov = "^2.11.1"
black = "^21.5b0"
isort = "^5.8.0"
flake8 = "^3.9.2"
bandit = "^1.7.0"
safety = "^1.10.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
